Imported prototype Elf
Imported prototype elf from Vee and integrate character movements. Cannot do animation due to no skeleton, but Vee is working on it. Also added line to player movement script to reset the players y position to 0 every update. I was getting some weird interactions that would cause the character float around. I could add a rigidbody and a collider because of the character controller. It was making everything super funky.