Updated Army Truck Recipe
Updated army truck recipe as it didnt like having two woods, it never updated the second wood.  Also the glue was bouncing on the left workstation so i removed the rigid body and modified the collider. It now works. The only issue now is that it wont spawn the gifts.